fix to prevent the creation of watermarks using alternative files, thanks to sherwood for finding this bug. If you use watermarks AND alternative files, you may have to recreate previews to correct for incorrect watermarked images.
git-svn-id: http://svn.resourcespace.org/svn/resourcespace@2137 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/image_processing.php
@@ -763,7 +763,7 @@ function create_previews_using_im($ref,$thumbonly=false,$extension="jpg",$previe
 				# echo $runcommand."<br>";
 				# Add a watermarked image too?
 				global $watermark;
-				if (isset($watermark) && ($ps[$n]["internal"]==1 || $ps[$n]["allow_preview"]==1))
+				if ($alternative==-1 && isset($watermark) && ($ps[$n]["internal"]==1 || $ps[$n]["allow_preview"]==1))
 					{
 					$path=get_resource_path($ref,true,$ps[$n]["id"],false,"",-1,1,true);
 					if (file_exists($path)) {unlink($path);}
